assignment_1/patient/pyproject.toml
[project]
name = "patient-api"
version = "0.1.0"
description = "환자 예약 서비스 API"
authors = [
    {name = "Developer", email = "developer@example.com"},
]
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.10",
    "cryptography>=46.0.3", 
    "fastapi[standard]>=0.121.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.44",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "coverage>=7.11.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]

[[tool.mypy.overrides]]
module = [
    "asyncmy.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E501", # Line too long
]

[tool.coverage.run]
concurrency = ["greenlet"]
